atspi_connection

Function set_session_accessibility

pub async fn set_session_accessibility(status: bool) -> Result<(), AtspiError>

Set the IsEnabled property in the session bus.

Assistive Technology provider applications (ATs) should set the accessibility IsEnabled status on the users session bus on startup as applications may monitor this property to enable their accessibility support dynamically.

See: The freedesktop - AT-SPI2 wiki

Example

    let result =  tokio_test::block_on( atspi_connection::set_session_accessibility(true) );
    assert!(result.is_ok());

Errors

when no connection with the session bus can be established,

if creation of a atspi_proxies::bus::StatusProxy fails

if the IsEnabled property cannot be read

the IsEnabled property cannot be set.