atspi_connection

Struct Peer

pub struct Peer { /* private fields */ }

Represents a peer with the name, path and connection for the P2P peer.

Implementations

impl Peer

pub fn unique_name(&self) -> &OwnedUniqueName

Returns the bus name of the peer.

pub fn well_known_name(&self) -> Option<&OwnedWellKnownName>

Returns the well-known bus name of the peer, if it has one.

pub fn socket_address(&self) -> &Address

Returns the socket Address of the peer.

pub fn connection(&self) -> &Connection

Returns the p2p Connection of the peer.

pub async fn try_from_bus_name( bus_name: BusName<'_>, conn: &Connection, ) -> AtspiResult<Self>

Try to create a new Peer from a bus name.

Errors

Returns an error if the application proxy cannot be created or if it does not support get_application_bus_address.
A non-existent bus name will also return an error.

pub async fn proxies(&self, path: &ObjectPath<'_>) -> AtspiResult<Proxies<'_>>

Returns a Proxies object for the given object path.
A Proxies object is used to obtain any of the proxies the object supports.

Errors

On invalid object path.

pub async fn as_root_accessible_proxy(&self) -> AtspiResult<AccessibleProxy<'_>>

Returns an AccessibleProxy for the root accessible object of the peer.

Errors

In case of an invalid connection.

pub async fn as_accessible_proxy( &self, obj: &ObjectRefOwned, ) -> AtspiResult<AccessibleProxy<'_>>

Returns an AccessibleProxy for the accessible object of the peer.

Errors

In case of an invalid connection or object path.